requirements.txt
numpy~=1.18.1
Pillow~=9.3.0
future~=0.18.2